[Ops] Drop o_pre activation, merge bwd pre-pass, reuse o as fwd scratch
Build on the bwd-fusion commit:

- **Don't save `o_pre` across fwd→bwd.** `b_o_pre = sum_l p_l * v_l` is
  recomputed inline in the bwd kernels from the saved `p` and `v` (which
  attnres already needs to keep around for the dv computation). Saves
  `[N, D]` of fp32 activation per attnres call. Aligns with the way
  Megatron-LM's `apply_attn_res` is wrapped in
  `activation_manager.checkpoint(..., enable=True)` — production never
  persists this state either, just there it's via outer checkpointing,
  here we make it cheap to do without the outer wrapper.

- **Merge the bwd pre-pass into pass 1.** Apply the layernorm-bwd-style
  factoring `do_pre = (wdy_o - xhat_o * c1_o) * o_rstd` so the per-l
  contributions to `b_dp` split into `b_dp[l] = o_rstd * (b_dp_w[l] -
  c1_o * b_dp_x[l])`, with `b_dp_w` / `b_dp_x` accumulating cleanly in
  the same D-loop that already loads `v` for `b_dp` / `b_z`. Drops a v
  load: bwd dv kernel goes from 3 v reads (pre-pass + pass 1 + pass 2)
  back down to 2 — the same as the no-onorm baseline. Net bwd v
  traffic vs the pre-fusion (save fp32 `o_pre`) version: −16% at
  L=15 fp16.

- **Reuse `o` as fwd scratch.** The fwd kernel needed an `[N, D]`
  scratch buffer to bridge pass 2 → pass 3 (compute `b_o`, then
  normalize after `o_rstd` is known). It's now allocated as the
  user-facing `o` itself, kept in fp32 inside the kernel under
  HAS_ONORM and cast to `v.dtype` once at the python-wrapper boundary.
  Saves an extra `[N, D]` fp32 fwd-time peak. `tl.debug_barrier()`
  between pass 2 and pass 3 ensures the staged values are visible
  before the third pass reads them back.

Naming and structural cleanups while in the file:

- Rename inner accumulators to follow fla's layernorm-bwd convention:
  `b_t1` / `b_t2` → `b_dp_w` / `b_dp_x` (matching `b_wdy_o = ow * do`),
  `b_a` / `b_b` collapsed into `b_wdy_o` / `b_xhat_o`. The dqdw kernel's
  `p_base` / `b_base` / `p_base_o` / `b_base_o` → `p_dqw` / `b_dqw` /
  `p_dow` / `b_dow` (match the tensor names they point at).
- Reorder `ctx.save_for_backward` and `fused_attnres_bwd` parameters
  to `(q, v, w, ow, p, rstd, o_rstd)` — fwd inputs followed by saved
  state, mirroring the bwd kernel's signature.
- `naive_attnres` runs fp32 end-to-end and casts to `residuals.dtype`
  once at the very end, matching the kernel's "fp32 internal, single
  final cast" path. Keeps the reference's reduction-order ULP noise as
  the only source of fp16 test diffs.

+++ fla/ops/attnres/naive.py
@@ -77,15 +77,19 @@ def naive_attnres(
         D = output_shape[-1]
         residuals = torch.stack(tuple(residual.view(-1, D) for residual in residuals), dim=0)
 
+    # all math runs in fp32 end-to-end; the final downcast back to
+    # `residuals.dtype` happens once, just before returning.
     v = residuals.float()
     k = F.rms_norm(v, (residuals.shape[-1],), rms_weight.flatten().float(), rms_eps)
     p = (einsum(k, query.flatten().float() * scale, "l ... d, d -> l ...")).softmax(dim=0)
-    o = einsum(p, v, "l ..., l ... d -> ... d").to(residuals.dtype)
+    o = einsum(p, v, "l ..., l ... d -> ... d")
     if output_shape is not None:
         o = o.view(output_shape)
 
     if output_rms_weight is not None:
-        o = F.rms_norm(o, (o.shape[-1],), output_rms_weight, rms_eps)
+        o = F.rms_norm(o, (o.shape[-1],), output_rms_weight.float(), rms_eps)
+
+    o = o.to(residuals.dtype)
 
     outputs = [o]
     if return_weights:
